added planet seed 01

+++ migrations/20170413101910_planets.js
@@ -3,8 +3,8 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('planet', (table) => {
     table.increments()
     table.string('name')
-    table.integer('value')
-    table.integer('distance')
+    table.double('value')
+    table.double('distance')
   })
 
 };
